Move test specs.
