+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "binary-stream"
-version = "5.1.0"
+version = "6.0.0"
 authors = [
   "Mathias Danielsen <mathiasda98@hotmail.com>",
   "Muji <muji@tmpfs.org>"
